Change formatting and generally sync up with Apple's GCC r150332.
llvm-svn: 55724

+++ llvm-gcc-4.2/gcc/testsuite/g++.dg/ext/align1.C
@@ -16,6 +16,7 @@ float f1 __attribute__ ((aligned));
 int
 main (void)
 {
-  printf ("%d %d\n", __alignof (a1), __alignof (f1));
+  /* APPLE LOCAL default to Wformat-security 5764921 */
+  printf ("%d %d\n", (int)__alignof (a1), (int)__alignof (f1));
   return (__alignof (a1) < __alignof (f1));
 }
